+++ test/ccxt_extract/schema_test.exs
@@ -91,7 +91,17 @@ defmodule CcxtExtract.SchemaTest do
       "handle_errors" => %{
         "method" => @sample_method_ast,
         "exceptions" => %{"broad" => %{"error" => "ExchangeError"}, "exact" => %{}},
-        "http_exceptions" => %{"429" => "RateLimitExceeded"}
+        "http_exceptions" => %{"429" => "RateLimitExceeded"},
+        "error_code_fields" => [
+          %{
+            "object" => "response",
+            "field" => "code",
+            "method" => "safeString",
+            "field2" => nil,
+            "roles" => ["error_code"],
+            "sentinel_values" => nil
+          }
+        ]
       },
       "parse_methods" => %{"parseTicker" => @sample_method_ast},
       "ws_methods" => %{"watchTicker" => @sample_method_ast},